requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
llvmlite==0.42.0
numba==0.59.0
pynndescent==0.5.11
umap-learn==0.5.5
hyperopt==0.2.7
python-dotenv==1.0.1
imbalanced-learn==0.10.1
scikit-learn==1.2.2
tqdm==4.65.0
mlflow==2.9.1
seaborn==0.12.2
skorch==0.12.1
cleanlab==2.5.0
plac==1.4.1
nni==3.0
pytorch-lightning==2.2.2